Reviews effects setting logic.
Stubs out the several different effects (currently return with "no
support" signal), annotates about the logic behind some of the
combinations, and adds logic to fail if an envelope is passed to an
effect that shouldn't have one.

Adds return status to VibrationController.cpp that can be used to infer
whether an effect could be applied.

Changes based in the documentation available at:
https://learn.microsoft.com/en-us/previous-versions/windows/desktop/ee416862(v=vs.85)
(DirectInput Interfaces reference)