Implements DISFFC_PAUSE and DISFFC_CONTINUE.
Now the SendForceFeedback() command will properly process a pause and
resume commands.

As I interpreted it, there's no need to actually stop the actuators,
just keep the effects in the exact spot they are -- and delay the end
for as long as the effects were paused. All this _should_ work, _should_
any software really use it.

Related GitHub issue: #11.