Reviews debug-mode logfile creation logic.
The logic had some issues, which prevented the log file from being
created even in a situation where it should have. The whole logic was
reviewed, and also simplified some code using preprocessor macros.

Also, for debug builds, ensures linear access to the log file, and that
not two parallel calls to LogMessage opens the file at the same time.